irrelevant fixme (2)
additional props set to true indicates accepting absolutely all and any additional properties. set to false to mean that additional properties are _not_ accepted - as in the server will reply 400/426 if such are present (that is of interest iff rweb acknowledges serde `deny_unknown_fields`).
it defaults to true.

additional props set to a schema indicates the shape that freeform properties must take. ex it is useful for `Map<String, T>` where string keys of the map become directly string keys of the sent JSON object.